Price the estimator in u64, not the host's pointer width
`POST /v1/estimate` returned `internal_error` for the top corner of the
contract's own provable envelope, and silently wrong numbers past it. Both
were live.

`estimate_params.rs` narrowed `lde_rows` with `as usize` before handing it to
`ResourceBoundedDft::estimate_scratch`, whose first act is
`height.checked_mul(width)`. That is a no-op on the native CLI and a
truncation in the wasm32 artifact that actually serves the endpoint, where
`usize` is 32 bits. Two regimes:

  * `rows * width >= 2^31` overflowed `checked_mul` and surfaced as
    `internal_error` / "generate_support_report". Reproduced against
    production at 2^24 rows x 256 width -- `MAX_ROWS` x `MAX_TRACE_WIDTH`,
    the best demonstration the product has. It now answers: conventional
    111,702,704,128 bytes against bounded 1,082,130,432, a 103x reduction.

  * `lde_rows >= 2^32` truncated toward zero, priced an empty matrix, and let
    `peak_resident_bytes` collapse onto the 64 MiB floor in `estimate_params`.
    At 2^31 rows x 64 that shipped 67,108,864 where the true figure is
    1,082,130,432 -- 16x understated, as a clean HTTP 200, with no error path
    anywhere that could have surfaced it. This regime is out-of-profile, which
    the estimator prices deliberately: pricing jobs you cannot run is the
    product, so "out of profile" is not "out of scope".

`estimate_scratch` now takes `u64` and narrows only the sub-FFT factors and
width, which are provably small, with `try_from` rather than `as`. The
module's own rule already said it: "a refusal is honest, a saturated number
that looks precise is not".

The parity gate had been proven to fail -- and still missed this, because both
its fixtures sat at or below 2^22 rows and 180 columns. Proving a gate CAN
fail is not proving it COVERS. Two fixtures now sit on either side of the
boundary; both were observed red against the previously committed wasm and
green after the rebuild.

Native output is unchanged: the two pre-existing fixtures still match the
OLD committed wasm byte-for-byte, which is what proves this is a no-op on
64-bit. Vendored wasm rebuilt via crates/hc-wasm/build.sh; the glue reproduces
byte-identically, so the compiled-in import name is untouched.

+++ crates/hc-plonky3/src/estimate_params.rs
@@ -249,15 +249,14 @@ pub fn estimate_from_params(
         .max(proof_checkpoint_peak);
 
     let dft = ResourceBoundedDft::new(policy.clone())?;
-    let trace_dft = dft.estimate_scratch(lde_rows as usize, width, false, field_bytes)?;
+    // `lde_rows` stays `u64` all the way into the DFT estimator. It was
+    // previously narrowed with `as usize` here, which is a no-op on the native
+    // CLI and a silent truncation in the wasm32 artifact that actually serves
+    // `POST /v1/estimate` — see the note on `ResourceBoundedDft::estimate_scratch`.
+    let trace_dft = dft.estimate_scratch(lde_rows, width as u64, false, field_bytes)?;
     // The quotient store holds one column per EXTENSION-field coefficient,
     // not a literal 2 — the same `extension_degree` derived above.
-    let quotient_dft = dft.estimate_scratch(
-        lde_rows as usize,
-        extension_degree as usize,
-        false,
-        field_bytes,
-    )?;
+    let quotient_dft = dft.estimate_scratch(lde_rows, extension_degree, false, field_bytes)?;
     let peak_resident_bytes = trace_dft
         .peak_resident_bytes
         .max(quotient_dft.peak_resident_bytes)
